+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/SuspendableNotifier.kt
@@ -6,12 +6,12 @@ import io.realm.kotlin.internal.interop.RealmInterop
 import io.realm.kotlin.internal.platform.freeze
 import io.realm.kotlin.internal.platform.runBlocking
 import io.realm.kotlin.internal.util.Validation.sdkError
+import io.realm.kotlin.internal.util.checkForBufferOverFlow
 import io.realm.kotlin.notifications.internal.Cancellable
 import kotlinx.atomicfu.AtomicRef
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.channels.BufferOverflow
-import kotlinx.coroutines.channels.ChannelResult
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.flow.Flow
@@ -105,7 +105,11 @@ internal class SuspendableNotifier(
                                 // Notifications need to be delivered with the version they where created on, otherwise
                                 // the fine-grained notification data might be out of sync.
                                 liveRef.emitFrozenUpdate(realm.snapshot, change, this@callbackFlow)
-                                    ?.let { checkResult(it)?.let { this@callbackFlow.close(it) } }
+                                    ?.run { // this: ChannelResult<T>
+                                        checkForBufferOverFlow()?.let { overflowException: Throwable ->
+                                            this@callbackFlow.close(overflowException)
+                                        }
+                                    }
                             }
                         }.freeze<io.realm.kotlin.internal.interop.Callback<RealmChangesPointer>>() // Freeze to allow cleaning up on another thread
                     val newToken =
@@ -135,17 +139,6 @@ internal class SuspendableNotifier(
             }
         }
     }
-
-    // Verify that notifications emitted to Streams are handled in an uniform manner
-    private fun checkResult(result: ChannelResult<Unit>): Throwable? =
-        if (!result.isClosed && result.isFailure) {
-            throw IllegalStateException(
-                "Cannot deliver object notifications. Increase dispatcher processing resources or" +
-                    " buffer the flow with buffer()"
-            )
-        } else {
-            null
-        }
 
     /**
      * Manually force a refresh of the Realm, moving it to the latest version.