+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/util/Channel.kt
@@ -17,6 +17,8 @@
 package io.realm.kotlin.internal.util
 
 import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.ChannelResult
+import kotlinx.coroutines.channels.SendChannel
 
 // Conventional try-with-resource wrapper for channels
 public inline fun <T : Channel<*>, R> T.use(block: (channel: T) -> R): R {
@@ -26,3 +28,17 @@ public inline fun <T : Channel<*>, R> T.use(block: (channel: T) -> R): R {
         this.close()
     }
 }
+
+// Public to be accessible from sync progress listeners
+public inline fun <T> SendChannel<T>.trySendCloseOnBufferOverflow(value: T): ChannelResult<Unit> =
+    trySend(value).apply { checkForBufferOverFlow()?.let { close(it) } }
+
+public inline fun <T> ChannelResult<T>.checkForBufferOverFlow(): Throwable? =
+    if (!isClosed && isFailure) {
+        IllegalStateException(
+            "Cannot deliver object notifications. Increase dispatcher processing resources or " +
+                "buffer the flow with buffer(...)"
+        )
+    } else {
+        null
+    }