+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/query/RealmResults.kt
@@ -24,6 +24,7 @@ import io.realm.kotlin.notifications.InitialResults
 import io.realm.kotlin.notifications.ResultsChange
 import io.realm.kotlin.notifications.UpdatedResults
 import io.realm.kotlin.types.BaseRealmObject
+import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
 
 /**
@@ -62,6 +63,9 @@ public interface RealmResults<T : BaseRealmObject> : List<T>, Deleteable, Versio
      *
      * The change calculations will on on the thread represented by
      * [Configuration.SharedBuilder.notificationDispatcher].
+     *
+     * The flow has an internal buffer of [Channel.BUFFERED] but if the consumer fails to consume the
+     * elements in a timely manner the flow will be completed with an [IllegalStateException].
      *
      * @return a flow representing changes to the RealmResults.
      */