+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/query/RealmSingleQuery.kt
@@ -19,6 +19,7 @@ package io.realm.kotlin.query
 import io.realm.kotlin.Deleteable
 import io.realm.kotlin.notifications.SingleQueryChange
 import io.realm.kotlin.types.BaseRealmObject
+import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
 
 /**
@@ -66,6 +67,9 @@ public interface RealmSingleQuery<T : BaseRealmObject> : Deleteable {
      * ```
      * The change calculations will run on the thread represented by
      * [RealmConfiguration.Builder.notificationDispatcher].
+     *
+     * The flow has an internal buffer of [Channel.BUFFERED] but if the consumer fails to consume the
+     * elements in a timely manner the flow will be completed with an [IllegalStateException].
      *
      * @return a flow representing changes to the [RealmObject] or [EmbeddedRealmObject] resulting from
      * running this query.